move interfaces to a shared lib
