+++ projects/adf-extensions/ng-package.json
@@ -3,6 +3,7 @@
   "dest": "../../dist/@alfresco/adf-extensions",
   "deleteDestPath": false,
   "lib": {
+    "languageLevel": ["dom", "es2017"],
     "entryFile": "src/public_api.ts"
   }
 }
