+++ projects/adf-extensions/ng-package.prod.json
@@ -2,6 +2,7 @@
   "$schema": "../../node_modules/ng-packagr/ng-package.schema.json",
   "dest": "../../dist/@alfresco/adf-extensions",
   "lib": {
+    "languageLevel": ["dom", "es2017"],
     "entryFile": "src/public_api.ts"
   }
 }
