+++ projects/adf-extensions/package.json
@@ -3,6 +3,8 @@
   "version": "0.0.1",
   "peerDependencies": {
     "@angular/common": "^6.0.0",
-    "@angular/core": "^6.0.0"
+    "@angular/core": "^6.0.0",
+    "@angular/http": "^6.1.4",
+    "alfresco-js-api": "^2.5.0"
   }
 }
