+++ projects/adf-extensions/src/lib/config/action.extensions.ts
@@ -1,0 +1,52 @@
+/*!
+ * @license
+ * Copyright 2016 - 2018 Alfresco Software, Ltd.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+export enum ContentActionType {
+    default = 'default',
+    button = 'button',
+    separator = 'separator',
+    menu = 'menu',
+    custom = 'custom'
+}
+
+export interface ContentActionRef {
+    id: string;
+    type: ContentActionType;
+
+    title?: string;
+    description?: string;
+    order?: number;
+    icon?: string;
+    disabled?: boolean;
+    children?: Array<ContentActionRef>;
+    component?: string;
+    actions?: {
+        click?: string;
+        [key: string]: string;
+    };
+    rules?: {
+        enabled?: string;
+        visible?: string;
+        [key: string]: string;
+    };
+}
+
+export interface ActionRef {
+    id: string;
+    type: string;
+    payload?: string;
+}
