+++ projects/adf-extensions/src/lib/store/states/profile.state.ts
@@ -15,25 +15,11 @@
  * limitations under the License.
  */
 
-import { NgModule, ModuleWithProviders } from '@angular/core';
-import { ExtensionLoaderService } from './services/extension-loader.service';
-
-@NgModule({
-    imports: [],
-    declarations: [],
-    exports: []
-})
-export class ExtensionsModule {
-    static forRoot(): ModuleWithProviders {
-        return {
-            ngModule: ExtensionsModule,
-            providers: [ExtensionLoaderService]
-        };
-    }
-
-    static forChild(): ModuleWithProviders {
-        return {
-            ngModule: ExtensionsModule
-        };
-    }
+export interface ProfileState {
+    id: string;
+    isAdmin: boolean;
+    firstName: string;
+    lastName: string;
+    userName?: string;
+    initials?: string;
 }
