+++ projects/adf-extensions/src/public_api.ts
@@ -1,5 +1,35 @@
-/*
- * Public API Surface of adf-extensions
+/*!
+ * @license
+ * Copyright 2016 - 2018 Alfresco Software, Ltd.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
  */
+
+export * from './lib/config/action.extensions';
+export * from './lib/config/extension-element';
+export * from './lib/config/extension-utils';
+export * from './lib/config/extension.config';
+export * from './lib/config/navbar.extensions';
+export * from './lib/config/permission.extensions';
+export * from './lib/config/routing.extensions';
+export * from './lib/config/rule.extensions';
+export * from './lib/config/sidebar.extensions';
+export * from './lib/config/viewer.extensions';
+
+export * from './lib/services/extension-loader.service';
+
+export * from './lib/store/states/navigation.state';
+export * from './lib/store/states/profile.state';
+export * from './lib/store/states/selection.state';
 
 export * from './lib/extensions.module';
