+++ src/app/components/current-user/current-user.component.ts
@@ -27,7 +27,8 @@ import { Component, ViewEncapsulation } from '@angular/core';
 import { Store } from '@ngrx/store';
 import { Observable } from 'rxjs';
 import { selectUser, appLanguagePicker } from '../../store/selectors/app.selectors';
-import { AppStore, ProfileState } from '../../store/states';
+import { AppStore } from '../../store/states';
+import { ProfileState } from '@alfresco/adf-extensions';
 
 @Component({
     selector: 'aca-current-user',
