+++ src/app/components/trashcan/trashcan.component.ts
@@ -30,9 +30,9 @@ import { PageComponent } from '../page.component';
 import { Store } from '@ngrx/store';
 import { selectUser } from '../../store/selectors/app.selectors';
 import { AppStore } from '../../store/states/app.state';
-import { ProfileState } from '../../store/states/profile.state';
 import { ExtensionService } from '../../extensions/extension.service';
 import { Observable } from 'rxjs';
+import { ProfileState } from '@alfresco/adf-extensions';
 
 @Component({
     templateUrl: './trashcan.component.html'
