+++ src/app/extensions/evaluators/core.evaluators.ts
@@ -23,7 +23,7 @@
  * along with Alfresco. If not, see <http://www.gnu.org/licenses/>.
  */
 
-import { RuleContext, RuleParameter } from '../rule.extensions';
+import { RuleContext, RuleParameter } from '@alfresco/adf-extensions';
 
 export function not(context: RuleContext, ...args: RuleParameter[]): boolean {
     if (!args || args.length === 0) {
