+++ src/app/extensions/evaluators/navigation.evaluators.ts
@@ -23,7 +23,7 @@
  * along with Alfresco. If not, see <http://www.gnu.org/licenses/>.
  */
 
-import { RuleContext, RuleParameter } from '../rule.extensions';
+import { RuleContext, RuleParameter } from '@alfresco/adf-extensions';
 
 export function isFavorites(
     context: RuleContext,
