--- src/app/extensions/rule.extensions.ts
@@ -1,50 +1,0 @@
-/*!
- * @license
- * Alfresco Example Content Application
- *
- * Copyright (C) 2005 - 2018 Alfresco Software Limited
- *
- * This file is part of the Alfresco Example Content Application.
- * If the software was purchased under a paid Alfresco license, the terms of
- * the paid license agreement will prevail.  Otherwise, the software is
- * provided under the following open source license terms:
- *
- * The Alfresco Example Content Application is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Lesser General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * The Alfresco Example Content Application is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Lesser General Public License for more details.
- *
- * You should have received a copy of the GNU Lesser General Public License
- * along with Alfresco. If not, see <http://www.gnu.org/licenses/>.
- */
-
-import { SelectionState } from '../store/states';
-import { NavigationState } from '../store/states/navigation.state';
-import { NodePermissions } from './permission.extensions';
-
-export type RuleEvaluator = (context: RuleContext, ...args: any[]) => boolean;
-
-export interface RuleContext {
-    selection: SelectionState;
-    navigation: NavigationState;
-    permissions: NodePermissions;
-
-    getEvaluator(key: string): RuleEvaluator;
-}
-
-export class RuleRef {
-    type: string;
-    id?: string;
-    parameters?: Array<RuleParameter>;
-}
-
-export interface RuleParameter {
-    type: string;
-    value: any;
-    parameters?: Array<RuleParameter>;
-}
