+++ src/app/testing/app-testing.module.ts
@@ -61,7 +61,7 @@ import { NodePermissionService } from '../services/node-permission.service';
 import { ContentApiService } from '../services/content-api.service';
 import { ExtensionService } from '../extensions/extension.service';
 import { ViewUtilService } from '../components/preview/view-util.service';
-import { ExtensionLoaderService } from '../extensions/extension-loader.service';
+import { ExtensionLoaderService } from '@alfresco/adf-extensions';
 
 @NgModule({
     imports: [
